Refactor Collage component
